refactor(test): extract signUpInWorkspaceAndGetAccessToken into shared util
Deduplicates the sign-up + mark-verified + get-token flow that was
inlined in both user.integration-spec and the onboarding integration
test.
